# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Panel de liderazgo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": ["usuarios"],
        "summary": "Users",
        "operationId": "users_api_usuarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "type": "array",
                  "title": "Response Users Api Usuarios Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["usuarios"],
        "summary": "Add",
        "operationId": "add_api_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{uid}": {
      "patch": {
        "tags": ["usuarios"],
        "summary": "Edit",
        "operationId": "edit_api_usuarios__uid__patch",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{uid}/deshabilitar": {
      "post": {
        "tags": ["usuarios"],
        "summary": "Disable",
        "operationId": "disable_api_usuarios__uid__deshabilitar_post",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{uid}/habilitar": {
      "post": {
        "tags": ["usuarios"],
        "summary": "Enable",
        "operationId": "enable_api_usuarios__uid__habilitar_post",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Role": {
        "type": "string",
        "enum": ["admin", "team_manager", "cliente"],
        "title": "Role"
      },
...
      "UserIn": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 12,
            "title": "Password"
          },
          "password_confirm": {
            "type": "string",
            "maxLength": 200,
            "title": "Password Confirm"
          }
        },
        "type": "object",
        "required": ["first_name", "last_name", "email", "role", "password", "password_confirm"],
        "title": "UserIn"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "first_name": {
            "type": "string",
            "title": "First Name",
            "readOnly": true
          },
          "last_name": {
            "type": "string",
            "title": "Last Name",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "active",
          "last_login_at",
          "created_at",
          "first_name",
          "last_name"
        ],
        "title": "UserOut"
      },
      "UserPatch": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "password_confirm": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Confirm"
          }
        },
        "type": "object",
        "required": ["first_name", "last_name", "email", "role"],
        "title": "UserPatch",
        "description": "Edición: la contraseña es opcional (vacía = no cambia)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}